requirements.txt
datasets==2.14.0
fastapi==0.101.1
httptools==0.6.0
hydra-core==1.3.2
konlpy==0.6.0
Korpora==0.2.0
prettytable==3.8.0
psycopg2-binary==2.9.7
scikit-learn==1.3.0
seaborn==0.12.2
SQLAlchemy==2.0.20
st-annotated-text==4.0.0
streamlit==1.26.0
transformers==4.31.0
uvicorn==0.23.2
wandb==0.15.8
httpcore==0.17.3
httpx==0.24.1
torch-model-archiver==0.8.1
torch-workflow-archiver==0.2.10
torchserve==0.8.1
--extra-index-url https://download.pytorch.org/whl/cu117
torch==2.0.1+cu117